significant simplification of instrumentation code; updated test structure related to #25
