--- hoopoe-core/src/main/java/hoopoe/core/instrumentation/EnterAdvice.java
@@ -1,11 +1,0 @@
-package hoopoe.core.instrumentation;
-
-import net.bytebuddy.asm.Advice;
-
-class EnterAdvice {
-
-    @Advice.OnMethodEnter
-    public static long before() {
-        return System.nanoTime();
-    }
-}
