+++ hoopoe-core/src/main/java/hoopoe/core/instrumentation/PluginRecorders.java
@@ -7,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.PARAMETER)
-@interface PluginActions {
+@interface PluginRecorders {
 }
