--- hoopoe-core/src/main/java/hoopoe/core/instrumentation/PluginsUnawareAdvice.java
@@ -1,24 +1,0 @@
-package hoopoe.core.instrumentation;
-
-import hoopoe.core.HoopoeProfilerFacade;
-import net.bytebuddy.asm.Advice;
-
-class PluginsUnawareAdvice {
-
-    @Advice.OnMethodExit(onThrowable = Throwable.class)
-    public static void after(@Advice.Enter long startTime,
-                             @Advice.Origin("#t") String className,
-                             @Advice.Origin("#m#s") String methodSignature,
-                             @MinimumTrackedTime long minimumTrackedTimeInNs) throws Exception {
-
-        if (HoopoeProfilerFacade.enabled && HoopoeProfilerFacade.profilingStartTime <= startTime) {
-            long endTime = System.nanoTime();
-
-            if (endTime - startTime >= minimumTrackedTimeInNs) {
-                HoopoeProfilerFacade.methodInvocationProfiler.profileMethodInvocation(
-                        startTime, endTime, className, methodSignature, 0, null, null, null
-                );
-            }
-        }
-    }
-}
