--- hoopoe-core/src/test/java/testies/JohnSnow.java
@@ -1,9 +1,0 @@
-package testies;
-
-public class JohnSnow {
-
-    public void goBehindTheWall() {
-        int i = 42;
-    }
-
-}
