+++ hoopoe-core/src/test/java/testies/PhilipJFryI.java
@@ -1,0 +1,9 @@
+package testies;
+
+public class PhilipJFryI {
+
+    public void beSpecial() {
+        // instrument me!
+    }
+
+}
