+++ hoopoe-core/src/test/java/testies/TurangaLeela.java
@@ -1,0 +1,11 @@
+package testies;
+
+public class TurangaLeela {
+
+    public static int yearBorn;
+
+    static {
+        // instrument me!
+    }
+
+}
